Refactor calendar package
