+++ packages/calendar/src/model/calendarEntry.test.ts
