+++ packages/calendar/src/model/calendarEntry.ts
