+++ packages/calendar/src/testing/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Nordeck IT + Consulting GmbH
+ * Copyright 2024 Nordeck IT + Consulting GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,5 @@
  * limitations under the License.
  */
 
-export * from './model';
-export * from './testing';
-export * from './utils';
+export { mockCalendar, mockCalendarEntry } from './testUtils';
+export { mockDateTimeFormatTimeZone } from './timezoneMockUtils';
