+++ packages/calendar/src/utils/calendarUtils/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Nordeck IT + Consulting GmbH
+ * Copyright 2024 Nordeck IT + Consulting GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,11 @@
  * limitations under the License.
  */
 
-export { generateRruleSet } from './calendarUtils/generateRruleSet';
-export { getCalendarEnd } from './calendarUtils/getCalendarEnd';
+export { generateRruleSet } from './generateRruleSet';
+export { getCalendarEnd } from './getCalendarEnd';
 export {
   isFiniteSeries,
   isRRuleEntry,
   isRRuleOverrideEntry,
   isSingleEntry,
-} from './calendarUtils/helpers';
-export { formatICalDate, parseICalDate, toISOString } from './dateTimeUtils';
-export { formatRRuleText, getOrdinalLabel, parseRRule } from './format';
-export {
-  isWeekdays,
-  normalizeByWeekday,
-  normalizeNumeric,
-  normalizeWeekday,
 } from './helpers';
